+++ eacDoojigger.php
@@ -21,9 +21,9 @@
  * Plugin URI:			https://eacDoojigger.earthasylum.com/
  * Update URI: 			https://swregistry.earthasylum.com/software-updates/eacdoojigger.json
  * Description:			{eac}Doojigger for WordPress - A new path to rapid plugin development. A powerful, extensible, multi-function architectural framework and utility plugin for WordPress.
- * Version:				3.0.4-RC1
+ * Version:				3.0.4-RC2
  * Requires at least:	5.8
- * Tested up to: 		6.7
+ * Tested up to: 		6.8
  * Requires PHP:		8.1
  * Author:				EarthAsylum Consulting
  * Author URI:			http://www.earthasylum.com